+++ extensions/mindsphere/mindsphere-http/src/main/java/com/siemens/mindsphere/datalake/edc/http/DataLakeClient.java
@@ -17,9 +17,9 @@ package com.siemens.mindsphere.datalake.edc.http;
 import java.net.URL;
 
 public interface DataLakeClient {
-    URL getDownloadUrl(String path) throws DataLakeException;
+    URL getPresignedUploadUrl(String path) throws DataLakeException;
 
-    URL getUploadUrl(String path) throws DataLakeException;
+    URL getPresignedDownloadUrl(String datalakePath) throws DataLakeException;
 
     boolean isPresent(String path) throws DataLakeException;
 }